feat: provide a way to override IPMI PXE boot method on `Server`
This adds a new field to the `Server` resource which can be used to
override default IPMI PXE boot method.

Fixes #987

(cherry picked from commit d8ef68b6b951df093d288c1dae6c0d32a9e00aec)
