+++ app/sidero-controller-manager/controllers/server_controller.go
@@ -33,8 +33,8 @@ import (
 	infrav1 "github.com/talos-systems/sidero/app/caps-controller-manager/api/v1alpha3"
 	metalv1alpha1 "github.com/talos-systems/sidero/app/sidero-controller-manager/api/v1alpha1"
 	"github.com/talos-systems/sidero/app/sidero-controller-manager/internal/power"
-	"github.com/talos-systems/sidero/app/sidero-controller-manager/internal/power/metal"
 	"github.com/talos-systems/sidero/app/sidero-controller-manager/pkg/constants"
+	siderotypes "github.com/talos-systems/sidero/app/sidero-controller-manager/pkg/types"
 )
 
 const (
@@ -50,7 +50,7 @@ type ServerReconciler struct {
 	Recorder  record.EventRecorder
 
 	RebootTimeout time.Duration
-	PXEMode       metal.PXEMode
+	PXEMode       siderotypes.PXEMode
 }
 
 // +kubebuilder:rbac:groups=metal.sidero.dev,resources=servers,verbs=get;list;watch;create;update;patch;delete
@@ -101,6 +101,11 @@ func (r *ServerReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 
 	if poweredOn {
 		s.Status.Power = "on"
+	}
+
+	pxeMode := r.PXEMode
+	if s.Spec.PXEMode != "" {
+		pxeMode = s.Spec.PXEMode
 	}
 
 	f := func(ready bool, result ctrl.Result) (ctrl.Result, error) {
@@ -204,7 +209,7 @@ func (r *ServerReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 
 		if !poweredOn {
 			// it's safe to set server to PXE boot even if it's already installed, as PXE server makes sure server is PXE booted only once
-			err = mgmtClient.SetPXE(r.PXEMode)
+			err = mgmtClient.SetPXE(pxeMode)
 			if err != nil {
 				log.Error(err, "failed to set PXE")
 				r.Recorder.Event(serverRef, corev1.EventTypeWarning, "Server Management", fmt.Sprintf("Failed to set to PXE boot once: %s.", err))
@@ -247,7 +252,7 @@ func (r *ServerReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctr
 			return f(false, ctrl.Result{RequeueAfter: constants.DefaultRequeueAfter})
 		}
 
-		err = mgmtClient.SetPXE(r.PXEMode)
+		err = mgmtClient.SetPXE(pxeMode)
 		if err != nil {
 			log.Error(err, "failed to set PXE")
 			r.Recorder.Event(serverRef, corev1.EventTypeWarning, "Server Management", fmt.Sprintf("Failed to set to PXE boot once: %s.", err))
