+++ app/sidero-controller-manager/pkg/types/types.go
@@ -12,3 +12,22 @@ const (
 	Boot404      BootFromDisk = "http-404"     // Return HTTP 404 response to iPXE.
 	BootSANDisk  BootFromDisk = "ipxe-sanboot" // Use iPXE script with `sanboot` command.
 )
+
+// PXEMode specifies PXE boot mode.
+type PXEMode string
+
+const (
+	PXEModeBIOS = "bios"
+	PXEModeUEFI = "uefi"
+)
+
+func (mode PXEMode) IsValid() bool {
+	switch mode {
+	case PXEModeBIOS:
+		return true
+	case PXEModeUEFI:
+		return true
+	default:
+		return false
+	}
+}
